# Softmax backward & Cross Entropy backward should both be optimal now