--- a/src/layers/CrossEntropyLoss.jl
+++ b/src/layers/CrossEntropyLoss.jl
@@ -30,11 +30,9 @@ function forward(l::CrossEntropyLoss, Y::Array{Float64,2}, label::Array{Float64,
   @assert size(Y) == size(label)
   @assert size(l.x) == size(Y)
 
-  l.x = log(Y)
-  l.x = -label.*l.x
-  # broadcast!(*, -l.x, label)
-  l.loss = sum(l.x,2)
   l.x = Y
+  l.loss = - sum(log(l.x) .* label,2)
+
 
   return l.loss, l.x
 end