--- a/src/layers/SoftMax.jl
+++ b/src/layers/SoftMax.jl
@@ -44,14 +44,9 @@ end
 
 function backward(l::SoftMax, DLDY::Array{Float64, 2}; kwargs...)
     # credits: https://stats.stackexchange.com/questions/79454/softmax-layer-in-a-neural-network?newreg=d1e89b443dd346ae8bccaf038a944221
-    m,n =size(l.x)
-    for batch=1:m
-      l.ly = l.y[batch,:]
-      l.jacobian .= -l.ly .* l.ly'
-      l.jacobian[diagind(l.jacobian)] .= l.ly.*(1.0.-l.ly)
-      # # n x 1 = n x n * n x 1
-      l.dldx[batch,:] = l.jacobian * DLDY[batch,:]
-    end
+
+    l.dldx = l.y .* (DLDY .- sum( l.y .* DLDY , 2))
+
     return l.dldx
 
 end