--- a/src/layers/SoftMax.jl
+++ b/src/layers/SoftMax.jl
@@ -33,11 +33,9 @@ end
 function forward(l::SoftMax, X::Array{Float64,2}; kwargs...)
 
     l.x = X
-    # iterating each row/picture
     l.lexp = exp(l.x)
     l.lsum = sum(l.lexp, 2)
-    # l.y = l.lexp./l.lsum
-    broadcast!(/, l.y, l.lexp, l.lsum)
+    l.y = l.lexp ./ l.lsum
     return l.y
 
 end