--- a/src/layers/SoftMaxCrossEntropy.jl
+++ b/src/layers/SoftMaxCrossEntropy.jl
@@ -1,14 +1,13 @@
 # include("LayerBase.jl")
 
 type SoftMaxCrossEntropyLoss <: LossCriteria
-    dldx   :: Array{Float64}
-    x      :: Array{Float64}
-    y      :: Array{Float64}
-    target :: Array{Float64}
-    loss   :: Array{Float64}
-    pred   :: Array{Int64}
-    label  :: Array{Int64}
-    iter   :: UnitRange{Int64}
+    dldx   :: Array{Float64} # backprop
+    x      :: Array{Float64} # input vector
+    y      :: Array{Float64} # output of softmax
+    exp    :: Array{Float64} # cache for exp(x)
+    lsum   :: Array{Float64} # cache for sum(exp,2)
+    loss   :: Array{Float64} # output of cross entropy loss
+    pred   :: Array{Int64}   # output for prediction
 
     function SoftMaxCrossEntropyLoss()
         return new(Float64[], Float64[], Float64[], Float64[], Float64[], Int64[], 1:1)
@@ -30,10 +29,9 @@ function init(l::SoftMaxCrossEntropyLoss, p::Union{Layer,Void}, config::Dict{Str
     l.x      = Array{Float64}(out_size)
     l.y      = Array{Float64}(out_size)
     l.loss   = Array{Float64}(N)
+    l.exp    = Array{Float64}(out_size)
     l.pred   = Array{Int64}(N)
-    l.label  = Array{Int64}(N)
-    l.iter   = 1:N
-    l.target = zeros(size(l.x))
+    l.lsum   = zeros(size(l.x))
 end
 
 function update(l::SoftMaxCrossEntropyLoss, input_size::Tuple;)
@@ -46,66 +44,28 @@ function update(l::SoftMaxCrossEntropyLoss, input_size::Tuple;)
     l.y      = Array{Float64}(input_size)
     l.loss   = Array{Float64}(N)
     l.pred   = Array{Int64}(N)
-    l.label  = Array{Int64}(N)
-    l.iter   = 1:N
-    l.target = zeros(size(l.x))
+    l.lsum = zeros(size(l.x))
 end
 
-function forward(l::SoftMaxCrossEntropyLoss, Y::Array{Float64,2}, label::Array{Int, 2}; kwargs...)
-    # TODO this is a questionable invariant; [label] ranges from 0 to 9 instead of being one hot
-    """
-    [label]  label[i] == 1 iff the data is classified to class i
-    [y]      final input to the loss layer
-    """
+function forward(l::SoftMaxCrossEntropyLoss, Y::Array{Float64,2}, label::Array{Float64, 2}; kwargs...)
     @assert size(Y, 2) == size(l.x, 2)
     local N = size(Y, 1)
     if N != size(l.x, 1)
-        update(l, size(Y))
+      update(l, size(Y))
     end
-    broadcast!(-, l.x, Y, maximum(Y, 2))
-    broadcast!(-, l.y, log(sum(exp(l.x),2)), l.x)
-    for i = 1:N
-        l.label[i] = label[i] + 1
-    end
-    # map!(x -> convert(Int64, x) + 1,    l.label, label)
+    l.x = Y
+    l.exp = exp(l.x)
+    l.lsum = sum(l.exp,2)
+    l.y = l.exp ./ l.lsum
 
-    for i = 1:N
-         l.loss[i] = l.y[i, l.label[i]]
-    end
-    # map!(i -> l.y[i, l.label[i]],       l.loss,  l.iter)
+    l.loss = - sum(log(l.y) .* label,2)
 
-    for i = 1:N
-        l.pred[i] = findmax(Y[i,:])[2] - 1
-    end
-    # map!(i -> findmax(Y[i,:])[2] - 1,   l.pred,  l.iter)
-
-    return l.loss, l.pred
+    return l.loss, l.x
 end
 
-function backward(l::SoftMaxCrossEntropyLoss, label::Array{Int, 2};kwargs...)
-    """
-    [label]  label[i] == 1 iff the data is classified to class i
-    [y]      final input to the loss layer
-    """
-    # TODO: [label] isn't used
-    # local N = size(l.x, 1)
-    # if size(l.target,1) != N
-    #     l.target    = zeros(size(l.x))
-    #     l.dldx = Array{Float64}(size(l.x))
-    # end
+function backward(l::SoftMaxCrossEntropyLoss, label::Array{Float64, 2};kwargs...)
 
-    # local TAR = zeros(size(l.x))
-    # for i = 1:N
-    fill!(l.target, 0)
-    for i = l.iter
-        l.target[i,l.label[i]] = 1
-    end
+    l.dldx = l.y .* sum(label,2) - label
 
-    # local Y = l.x
-    # Y = exp(broadcast(+, Y, - maximum(Y,2)))
-    l.dldx = exp(l.x)
-    broadcast!(/, l.dldx, l.dldx, sum(l.dldx,2))
-    broadcast!(-, l.dldx, l.dldx, l.target)
-    # return Y .- TAR
     return l.dldx
 end