Resolves #7