@fufjvnvnf 2017-07-15
Resolves #7