@stevenygd 2017-07-15
I wouldn't say it resolves #7 . Since the benchmark shows that the improvement is quite marginal:

# Before:
```
Old forward uses on average 0.008957189829999998 seconds
New forward uses on average 0.010831998020000002 seconds
Old backward uses on average 0.005899088019999998 seconds
New backward uses on average 0.013368990589999996 seconds
Old forward allocates on average 4.24580416e6 bytes
New forward allocates on average 8.3952e6 bytes
Old backward allocates on average 2.10334016e6 bytes
New backward allocates on average 8.3952e6 bytes
```

# After
```
Old forward uses on average 0.009047210220000003 seconds
New forward uses on average 0.011174593820000002 seconds
Old backward uses on average 0.00592484136 seconds
New backward uses on average 0.013712984500000004 seconds
Old forward allocates on average 4.24579888e6 bytes
New forward allocates on average 8.3952e6 bytes
Old backward allocates on average 2.10333488e6 bytes
New backward allocates on average 8.3952e6 bytes
```