@fufjvnvnf 2017-07-15
Both backwards are now only one line with no loop. I will later refactor the original `SoftmaxCrossEntropyLoss` layer, whose speed is likely to decrease under the new float matrix interface. When we compare the benchmark of the two layers under the same interface we should have a clearer expectation for the speed of these layers.