@stevenygd 2017-07-15
@fufjvnvnf I'm not talking about the comparison between new layers and the old ones. The performance of the new layers stay the same before and after this PR.