@fufjvnvnf 2017-07-16
Closing this PR but keeping the issue open. I'll also keep these changes in my next PR which modifies the interface of `SoftmaxCrossEntropy` layer since it still increases the speed slightly and the code looks cleaner & nicer.